added some ui basics and ability to start and stop sessions

+++ src/HeartWave-Project/HistoryManager/HistoryManager.cpp
@@ -76,8 +76,8 @@ bool HistoryManager::readRecordsFromFile() {
         float medPercentage;
         float highPercentage;
         float avgCoherance;
-        int lengthOfSession;
-        int achievementScore;
+        float lengthOfSession;
+        float achievementScore;
 
         stream >> id;
         stream >> time;
